+++ bin/actions/0220022-74.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+#: ◣ + Numpad ""-"" - Pause/Play audio
+
+xdotool key XF86AudioPlay