+++ CHANGELOG.md
@@ -3,8 +3,9 @@ xmchord Changelog
 
 v0.4.7 - Not tagged yet
 -----------------------
+* Add Toggle play/pause audio via: ◣ + Numpad "-"
 * Add workspace actions: ◣ + 0 / ◣ + 1, ◣ + 2, ◣ + 3, ◣ + 4 - Move active window 
-  to resp. worksapce
+  to resp. workspace
 * Add Mattermost action: ◣ + W - Type ":wave:"
 * Add Mattermost action: ◣ + 1 - Type ":+1:" (= thumbs-up icon)